* develop: (648 commits)
  ONM-2109: Add from date support on export contents command
  ONM-159: Improve placeholder text on manager image size
  Fix template and controller for settings
  ONM-2090 Reset opcache with tmp/restart file presence
  ONM-2108: Check for more than one newsItem on EP NewsML parse
  OC-1944: Add setTargeting with module value for DFP tags
  ONM-2102: Replace required="required" properties in backend forms
  ONM-2100: Improve algorithm of Instance Sync delete
  ONM-2101: Fix static pages link in advanced search template
  ONM-2100: Improve algorithm of instance sync delete action
  ONM-873: Check if layout is valid when going into the frontpage manager
  ONM-2022: Fix logout arguments method in SecurityListener
  ONM-2099: Disallow to create users with duplicated email/username from backend
  Fix warning on PHP 7.1 on ONM/SimpleMenu
  ONM-2098: Fix usage of core:status:check service in /framework
  ONM-2095: Add zamora24horas.com host as exclusion for iframe in ads
  ONM-2094: Remove UTF8 C0 control chars from StringUtils::generateSlug
  ONM-2088: Fix content uri link in comments backend listing
  Reorganize items in sidebar for manager
  Add missing blank line
  ...

+++ src/BackendWebService/Controller/ContentController.php
@@ -11,7 +11,7 @@
 namespace BackendWebService\Controller;
 
 use Common\Core\Annotation\Security;
-use Onm\Framework\Controller\Controller;
+use Common\Core\Controller\Controller;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 
@@ -44,14 +44,6 @@ class ContentController extends Controller
         $results = $em->findBy($search, $order, $elementsPerPage, $page);
         $results = \Onm\StringUtils::convertToUtf8($results);
         $total   = $em->countBy($search);
-
-        foreach ($results as &$result) {
-            $createdTime = new \DateTime($result->created);
-            $result->created = $createdTime->format(\DateTime::ISO8601);
-
-            $updatedTime = new \DateTime($result->changed);
-            $result->changed = $updatedTime->format(\DateTime::ISO8601);
-        }
 
         return new JsonResponse(
             array(
@@ -87,14 +79,6 @@ class ContentController extends Controller
         $results = \Onm\StringUtils::convertToUtf8($results);
         $total   = $em->countBy($search);
 
-        foreach ($results as &$result) {
-            $createdTime = new \DateTime($result->created);
-            $result->created = $createdTime->format(\DateTime::ISO8601);
-
-            $updatedTime = new \DateTime($result->changed);
-            $result->changed = $updatedTime->format(\DateTime::ISO8601);
-        }
-
         return new JsonResponse(
             array(
                 'extra'             => $this->loadExtraData($results),
@@ -124,12 +108,21 @@ class ContentController extends Controller
 
         if (!is_null($content->id)) {
             try {
-                $content->delete($id);
-                $success[] = array(
-                    'id'      => $id,
-                    'message' => _('Item deleted successfully'),
-                    'type'    => 'success'
-                );
+                if (!$this->get('core.security')->hasPermission('CONTENT_OTHER_UPDATE')
+                    && !$content->isOwner($this->getUser()->id)
+                ) {
+                    $errors[] = [
+                        'error',
+                        _("You don't have enough privileges for modify this content.")
+                    ];
+                } else {
+                    $content->delete($id);
+                    $success[] = array(
+                        'id'      => $id,
+                        'message' => _('Item deleted successfully'),
+                        'type'    => 'success'
+                    );
+                }
             } catch (Exception $e) {
                 $errors[] = array(
                     'id'      => $id,
@@ -176,8 +169,20 @@ class ContentController extends Controller
 
                 if (!is_null($content->id)) {
                     try {
-                        $content->delete($id);
-                        $updated[] = $id;
+                        if (!$this->get('core.security')->hasPermission('CONTENT_OTHER_UPDATE')
+                            && !$content->isOwner($this->getUser()->id)
+                        ) {
+                            $errors[] = [
+                                'error',
+                                sprintf(
+                                    _("You don't have enough privileges to delete the content with id %s."),
+                                    $content->id
+                                )
+                            ];
+                        } else {
+                            $content->delete($id);
+                            $updated[] = $id;
+                        }
                     } catch (Exception $e) {
                         $errors[] = array(
                             'id'      => $id,
